+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   - Add optional `stopOnError` parameter to `lexAll`.
   - Change `Token` and `Definition` from type to interface.
   - If token has no error, the error field will be `undefined`.
+- Parser
+  - If ASTNode has no error, the error field will be `undefined`.
+  - LR Parser
+    - If ReducerContext has no error, the error field will be `undefined`.
 - Optimize document / comments.
 - Optimize code.
 - More test code.